feat: threaded FileSystem over a DataSource abstraction (TODO_4 chunk 4a)
Replace the placeholder FileSystem with a real, threaded directory browser:
a DataSource interface + LocalDataSource, a worker-thread scan with the
main-thread swap contract, virtual root listing the sources, and
requestFile/consumeFetchResult resolution. FileEntry gains byte size + type.

Route all playback through FileSystem::requestFile so remote sources can later
resolve asynchronously; Application keeps a retry cursor and consumes the
resolved FetchResult in update(). Wire one LocalDataSource, the start path, and
the destroy-before-player ordering in main.cpp.

+++ src/gui/Gui.cpp
@@ -257,7 +257,7 @@ void Gui::drawFileBrowser(const std::vector<FileEntry> &files, const std::functi
                         onFileClick(file_entry);
                     }
                     ImGui::TableNextColumn();
-                    ImGui::Text("%i Kb", file_entry.file_size);
+                    ImGui::Text("%lld Kb", static_cast<long long>(file_entry.file_size));
                 }
             }
         }
